+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Require users to be logged in to be able to add a profession
 - Change roles to permissions
 - Update entities on seed, rather than replacing them
+- Take users from internal organisations page to internal profession page via link, rather than to public-facing page
 
 ## [release-001] - 2022-01-10